requirements.txt
docopt==0.6.2
Pillow==8.2.0
pkg-resources==0.0.0
